+++ js/util.js
@@ -1,0 +1,14 @@
+function getRandom(min, max) {
+  return (min >= 0 && max >= 0) ? Math.floor(Math.random() * (max - min) + min) : NaN;
+}
+function checkLengthString(str, strLength) {
+  return str.length <= strLength;
+}
+function getNumber(arr) {
+  const arrElement = arr[0];
+  arr.splice(0, 1);
+  return arrElement;
+}
+checkLengthString(1, 2);
+const getRandomArrayElement = (elements) => elements[getRandom(0, elements.length - 1)];
+export { getRandom, getRandomArrayElement, getNumber };
